LPS-98167 move OAuth2State
