+++ modules/apps/document-library-opener/document-library-opener-api/bnd.bnd
@@ -5,5 +5,6 @@ Export-Package:\
 	com.liferay.document.library.opener.constants,\
 	com.liferay.document.library.opener.exception,\
 	com.liferay.document.library.opener.model,\
+	com.liferay.document.library.opener.oauth2,\
 	com.liferay.document.library.opener.service,\
 	com.liferay.document.library.opener.service.persistence
