+++ modules/apps/document-library-opener/document-library-opener-api/src/main/java/com/liferay/document/library/opener/oauth2/OAuth2State.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.document.library.opener.model;
+package com.liferay.document.library.opener.oauth2;
 
 import java.io.Serializable;
